backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
Deprecated==1.3.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.6
google-ai-generativelanguage==0.6.10
google-api-core==2.30.0
google-api-python-client==2.154.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.3
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
limits==5.8.0
lxml==6.0.2
numpy==2.2.1
openpyxl==3.1.5
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
pillow==12.1.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.3.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
PyYAML==6.0.3
requests==2.33.0
rsa==4.9.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
tenacity==9.1.4
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
xlsxwriter==3.2.9
